feat:hrinka２さんのPARTとか諸々取り込みOK
