+++ command/KICK.cpp
